packages/protocols/x402-ampersend-sdk/python/ampersend-sdk/pyproject.toml
[project]
name = "ampersend-sdk"
version = "0.0.7"
description = "Python SDK for ampersend"
readme = "README.md"
dependencies = [
    "a2a-sdk",
    "google-adk>=1.14.0,<=1.14.1",
    "x402-a2a",
    "x402>=0.1.4",
    "httpx>=0.27.0",
    "siwe>=4.0.0",
    "eth-account>=0.11.0",
    "eth-utils>=2.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]
